Waveform rendering shader: Group write to termination flag
Instead of writing to the termination flag from different
point in the code, which might lead to some writes being
issued while there are already pending writes from other
threads coming from other code path.

This seems to be an issue on intel at least.

So instead each thread keeps a local flag and then in
the common execution path, that flag might trigger a write
to the shared variable and immediately after a barrier will
be issued before any work item uses the flag or try to issue
another write to it.

Fixes #325

Signed-off-by: Sylvain Munaut <tnt@246tNt.com>